Dealing with difficulties in setting the context
On my Mac, desktop-linux kept coming up as the default.  Not sure why.

ISET3d-tiny uses 'default' as the default.

That broke everything.  I added piDockerCurrentContext so that we set the current context.  That seems to work (again) for my in local mode and in remote mode on orange.

Puzzled why things worked for me in the past.  But they did.